ai-ops-desk/requirements.txt
langgraph>=0.2.0
openai>=1.0.0
pymongo>=4.7.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
httpx>=0.25.0
uvloop>=0.17.0
typing-extensions>=4.5.0
galileo>=1.17.1